+++ axelor-base/src/main/java/com/axelor/apps/base/web/weeklyplanning/WeeklyPlanningController.java
@@ -21,24 +21,22 @@ import com.axelor.apps.base.db.WeeklyPlanning;
 import com.axelor.apps.base.service.weeklyplanning.WeeklyPlanningService;
 import com.axelor.exception.AxelorException;
 import com.axelor.exception.service.TraceBackService;
+import com.axelor.inject.Beans;
 import com.axelor.rpc.ActionRequest;
 import com.axelor.rpc.ActionResponse;
-import com.google.inject.Inject;
 
 public class WeeklyPlanningController {
 
-  @Inject private WeeklyPlanningService weeklyPlanningService;
-
   public void initPlanning(ActionRequest request, ActionResponse response) {
     WeeklyPlanning planning = request.getContext().asType(WeeklyPlanning.class);
-    planning = weeklyPlanningService.initPlanning(planning);
+    planning = Beans.get(WeeklyPlanningService.class).initPlanning(planning);
     response.setValue("weekDays", planning.getWeekDays());
   }
 
   public void checkPlanning(ActionRequest request, ActionResponse response) {
     WeeklyPlanning planning = request.getContext().asType(WeeklyPlanning.class);
     try {
-      planning = weeklyPlanningService.checkPlanning(planning);
+      planning = Beans.get(WeeklyPlanningService.class).checkPlanning(planning);
     } catch (AxelorException e) {
       TraceBackService.trace(response, e);
     }